Fix link to serializable set of props
The "Passing props from Server to Client Components" section currently links to the set of argument types that are serializable when calling a Server Action. The list of serializable argument types that can be passed to a Server Action is different than the list of serializable prop types that can be passed from Server to Client Components.

The updated link will new correctly link to the list of serializable prop types, which is what the Next.js docs section is talking about.